pyproject.toml
[project]
name = "TasKanLine"
version = "0.1.0"
description = "Add your description here"
author = ["Yernaz Kenesbekov <yernazkenesbekov24@gmail.com>"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "asyncpg>=0.31.0",
    "aiosqlite>=0.21.0",
    "argon2-cffi>=25.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
    "ty>=0.0.1a28",
]
